ids to long, vat as enum value

+++ src/main/java/pl/futurecollars/invoicing/db/jpa/InvoiceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import pl.futurecollars.invoicing.model.Invoice;
 
 @Repository
-public interface InvoiceRepository extends CrudRepository<Invoice, Integer> {
+public interface InvoiceRepository extends CrudRepository<Invoice, Long> {
 
 }
